Construct database
